+++ edison-validation/src/main/java/de/otto/edison/validation/validators/IsEnum.java
@@ -9,7 +9,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Documented
-@Constraint(validatedBy = {EnumValidator.class, EnumSetValidator.class})
+@Constraint(validatedBy = {EnumValidator.class, EnumListValidator.class})
 @Target({ElementType.METHOD, ElementType.FIELD, ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface IsEnum {