+++ edison-validation/src/test/java/de/otto/edison/validation/validators/EnumListValidatorIntegrationTest.java
@@ -8,24 +8,23 @@ import org.springframework.boot.test.context.SpringBootTest;
 import javax.validation.ConstraintViolation;
 import javax.validation.Validator;
 import java.util.Arrays;
-import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
+
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 
 @SpringBootTest(classes = ValidationConfiguration.class)
-public class EnumSetValidatorIntegrationTest {
+public class EnumListValidatorIntegrationTest {
 
     @Autowired
     private Validator validator;
 
     @Test
     public void shouldReturnCorrectConstraintViolationMessage() {
-
         //given
-
-        TestClass testClass = new TestClass(createSet("foo","bar","nofoo"));
+        TestClass testClass = new TestClass(Arrays.asList("foo", "bar", "nofoo"));
 
         //when
         Set<ConstraintViolation<TestClass>> constraintViolations = validator.validate(testClass);
@@ -37,10 +36,8 @@ public class EnumSetValidatorIntegrationTest {
 
     @Test
     public void shouldReturnCorrectConstraintViolationMessageWithMultipleValues() {
-
         //given
-
-        TestClass testClass = new TestClass(createSet("foo","nobar","nofoo"));
+        TestClass testClass = new TestClass(Arrays.asList("foo", "nobar", "nofoo"));
 
         //when
         Set<ConstraintViolation<TestClass>> constraintViolations = validator.validate(testClass);
@@ -48,13 +45,6 @@ public class EnumSetValidatorIntegrationTest {
         //then
         assertThat(constraintViolations, hasSize(1));
         assertThat(constraintViolations.iterator().next().getMessage(), is("Unbekannte Enum-Werte: nobar,nofoo."));
-
-    }
-
-    private Set<String> createSet(String... values) {
-        return new HashSet<String>() {{
-            addAll(Arrays.asList(values));
-        }};
     }
 
     enum TestEnum {
@@ -63,10 +53,10 @@ public class EnumSetValidatorIntegrationTest {
     }
 
     private class TestClass {
-        @IsEnum(enumClass = EnumSetValidatorTest.TestEnum.class)
-        private Set<String> someEnums;
+        @IsEnum(enumClass = TestEnum.class)
+        private List<String> someEnums;
 
-        public TestClass(Set<String> someEnums) {
+        public TestClass(List<String> someEnums) {
             this.someEnums = someEnums;
         }
     }